feat(bzlmod): Implementing wheel annotations via whl_mods (#1278)
This commit implements a bzlmod extension that allows users to create
"annotations" for wheel builds. The wheel_builder.py accepts a JSON file
via a parameter called annotations; this extension creates those JSON
files. The pip extension accepts a Label -> String dict argument of the
JSON files.

This feature is renamed to `whl_mods` because the JSON files are handled
differently
and the name "annotations" is uninformative. This modifies the creation
of the BUILD
files and their content, and is much more than just adding some notes
about a whl.

The whl_mod extension wheel names and the wheel names in pip must match.

Closes: https://github.com/bazelbuild/rules_python/issues/1213