feat(bzlmod): Use a common constant for detecting bzlmod being enabled
Various parts of the codebase detect whether bzlmod is enabled or not.
Most of them copy/paste the same if @ in Label(..) trick, and use a
comment to explain what they're doing.

Rather that copy/paste that everywhere, this commit uses a constant defined
that does this once, then code can simply check "if BZLMOD_ENABLED".

This commit also refactors the BZLMOD_ENABLED variable to its own file,
so that we are not loading skylib all of the time.