feat(bzlmod): Use a common constant for detecting bzlmod being enabled (#1302)
Various parts of the codebase detect whether bzlmod is enabled or not.
Most of them copy/paste the same if @ in Label(..) trick and use a
comment to explain what they're doing.

Rather than copy/paste that everywhere, this commit uses a constant
defined that does this once and reuses the constant value to determine
if bzlmod is enabled.

Closes: https://github.com/bazelbuild/rules_python/issues/1295