Allow bzlmod pip.parse to reference the default python
This commit defaults the python_version attribute to the default
version of Python.  This allows a user to use the version we
define in rules_python MODULE.bazel.  Also, a user can now
not set the python_version attribute on a pip tag_class
and we automatically set the interpreter.