fix(bzlmod)!: Changing repository name "python_aliases" to "python_versions" (#1304)
I think this name is more informative for a public API. The
functionality it exposes are rules/macros that use a specific Python
version to be used. These aren't really aliases.

This commit renames "python_aliases" to "python_versions". This isn't
technically a breaking
change because bzlmod support is still beta, but we'll flag it as such
just in case.

BREAKING CHANGE:
* The `python_aliases` repo is renamed to `python_versions`. You will
need to either
update references from `@python_aliases` to `@python_versions`, or use
repo-remapping
to alias the old name (`use_repo(python,
python_aliases="python_versions")`)

Closes #1273