fix(multi-versions): correctly default 'main' arg for transition rules (#1316)
This fixes a bug where the version-aware rules required `main` to always
be explicitly
specified. This was necessary because the main file is named after the
outer target
(e.g. "foo"), but usage of the main file is done by the inner target
("_foo"). The net
effect is the inner target looks for "_foo.py", while only "foo.py" is
in srcs.

To fix, the wrappers set main, if it isn't already set, to their name +
".py"

Work towards #1262