revert(bzlmod)!: allow bzlmod pip.parse to implicitly use default python version (#1341)
Reverts bazelbuild/rules_python#1303

The main issue is that `pip.parse()` accepts a locked requirements file
-- this means
the requirements are specific to a particular Python version[1]. Because
the default Python
version can arbitrarily change, the lock file may not be valid for the
Python version
that is used at runtime. The net result is a module will use
dependencies for e.g. Python
3.8, but will use 3.9 at runtime. Additionally, the dependencies
resolved for 3.8 will
be created under names such as `@foo_39` (because that's the
python_version pip.parse sees),
which is just more confusing.

BREAKING CHANGE:
* pip.parse() must have `python_version` explicitly set. Set it to the
Python version
    used to resolve the requirements file.

[1] Lock files aren't necessarily version specific, but we don't
currently support the
environment markers in lock files to make them cross-python-version
compatible.