fix: Don't require default Python version for pip hubs (#1344)
This fixes the issue where a sub-module was required to always have a
pip.parse() call configured for the default Python version if it used
any pip.parse() call. Such a
requirement puts sub-modules in an impossible situation:
  * If they don't have the default version, they'll get an error.
* If they register the default version, but also register a specific
version,
they'll potentially cause an error if a root module changes the default
to
match their specific version (because two pip.parse() calls for the same
version
    are made, which is an error).

The requirement to have the default version registered for a pip hub was
only present to satisfy the `whl_library_alias` repository rule, which
needed a Python version to map `//conditions:default` to.

To fix, the `whl_library_alias` rule's `default_version` arg is made
optional. When None is passed, the `//conditions:default` condition is
replaced with a `no_match_error` setting. This prevents the pip hub from
being used with the version-unaware rules, but that makes sense: no
wheels were setup for that version, so it's not like there is something
that
can be used anyways.

Fixes #1320