feat! add entry_point macro to rules_python
Introduce a new `entry_point` macro to `rules_python` as opposed to the
`hub` repository which allows users to generate an `entry_point` script
for a given package. This will check the `console_scripts` key in the
`entry_points.txt` dist-info file and avoids eager fetching of third
party repositories because it is a `genrule` and a `py_binary`
underneath the hood and exists in `rules_python`.

This is a breaking change for bzlmod users as they will have to start
using `@rules_python//python:entry_point.bzl`. For others this new macro
is available to be used, but the old code is still present.

Fixes #1362
Fixes #543
Fixes #979
Fixes #1262
Closes #980
Closes #1294