fix: Fixing select when we don't have specific platforms
In certain releases of Python we do not have various platforms like
ppc. When using the call PLATFORMS.keys() we get all of the
platforms instead of the platforms that are included with the Python
version.

We get the error "no such package" because, for instance,
python_3_10_ppc64le-unknown-linux-gnu does not exist in Python
3.10.9 (see versions.bzl).

You can recreate this error by running a bazel query command inside
the examples/bzlmod directory. If you run:

bazel query "deps(@python_3_10//:python_headers)"

With a version of the rules previous to this commit, you will get
an error. With this commit, you get a printout of all of the
python_headers for all of the platforms that are supported in the
Python 3.10 version used by the repository.

Instead of using PLATFORMS.keys(), we use the sha256 keys in
TOOL_VERSIONS dict based on the full Python version. This change
now returns only the platforms supported in the Python version.

Fixes: #1326