fix: bcr releaser email (#1392)
Fixes https://github.com/bazelbuild/bazel-central-registry/pull/863.

The aspect email is no longer associated with the github user, so the
CLA bot but doesn't
think think the CLA is signed. To fix, change the email the BCR PRs are
published under
to an address that is associated with the github user (and thus the
CLA).