feat(pip_parse): support patching 'whl_library'
Before that the users had to rely on patching the actual wheel files and
uploading them as different versions to internal artifact stores if they
needed to modify the wheel dependencies. This is very common when
breaking dependency cycles in `pytorch` or `apache-airflow` packages.
With this feature we can support patching external PyPI dependencies via
pip.whl_override tag class to fix package dependencies and/or a broken
`RECORD` metadata file.

add the wheelmaker to the tool list that is used by whl_library

use whlmaker to repack a wheel