rm(python_repositories): remove unused get_interpreter_dirname
This is not used anymore in our `whl_library`, so remove it.

NOTE: if we think that there are uses outside our repo, then we should
reconsider, but given that it is very trivial for the user to write the
same helper, I would be inclined to reduce the API scope.