Import pycross_wheel_library
This patch imports a few files from jvolkman/rules_pycross at
757033ff8afeb5f7090b1320759f6f03d9c4615c.

I would like to re-use this rule for the `pypi_install` repo rule that
I'm working on. This rule extracts a downloaded wheel and generates an
appropriate `PyInfo` provider for it.

All the non-BUILD files are taken as-is without modification. A
followup patch will make tweaks so that the code can be used from
within rules_python.

References: #1360