fix: don't set distribs in version transitioning rule (#1412)
This makes the version-aware transition rule compatible with an upcoming
Bazel change that
disallows setting unknown attributes to None (the `distribs` attribute,
in this case).

The `distribs` attribute was common to all rules, but it has been long
deprecated and
it won't be part of every rule in upcoming Bazel versions. The previous
implementation
resulted in setting `distribs = None` on the target. Bazel won't support
setting
undefined attributes to None.

Addresses: https://github.com/bazelbuild/bazel/issues/19403

---------

Co-authored-by: Richard Levasseur <rlevasseur@google.com>