Rename `pycross_wheel_library` to `py_wheel_library` and make it work
This patch changes the name of the rule to reflect the fact that it's
not exactly the same as the one in rules_pycross.

I also took this opportunity to delete the superfluous
`namespace_pkgs.py` library (plus test) since we have a nearly
identical version already in the main repo.

I added a test to validate that the rule functions at a basic level.

References: #1360