feat, refactor(pystar): bzl_library for packaging.bzl; fix pystar doc building and py_wheel
Changed `py_wheel` to load `py_binary` instead of using
`native.py_binary`. This caused the doc generation to fail because
of the additional loads(), so the doc libraries were refactored to
represent the additional loads.

Docs were failing to build because Stardoc wasn't able to process
the value `cc_helper.use_cpp_toolchains()` returned. For some reason,
under Stardoc, the value is some mocked out value that it can't handle.
This is easily fixed by just using the regular way for referencing an
optional toolchain; the labels the two use are the same.

Work towards #1069