tests: Run default tests with Starlark implementation of rules.
Since Bazel 7 isn't yet available, the rolling release of Bazel is
used. This means there's about 2 weeks between when changes in Bazel
become available (similar to within Google).

Only the default tests (i.e. not examples and integrationt tests) are
run on Ubuntu because we don't have capacity to run a more comprehensive
matrix. When Bazel 7 is released, we'll be able to update our coverage
to be more comprehensive (dropping 5.4 support frees up capacity).

Work toward 1069