chore: bump internal_deps (#1322)
Before this PR the dependencies were out of date and the
`pyproject_hooks` was
missing allowing `pip-compile` to use `pyproject.toml` instead of
`requirements.in`. With the update we are more likely to support the new
features but they are not tested in this PR as the main goal is to
update all of
the files that need regenerating due to upgraded `pip-tools`.

Here we:
* Upgrade all of the deps to the latest versions.
* Update all of the requirements and gazelle manifest files.
* Add a quirk to ensure fix `pip_compile` on Windows.

Summary of version changes:
* build: 0.9.0 -> 0.10.0
* click: 8.0.1 -> 8.1.6
* importlib_metadata: 1.4.0 -> 6.8.0
* more_itertools: 8.13.0 -> 10.1.0
* packaging: 22.0 -> 23.1
* pip: 22.3.1 -> 23.2.1
* pip_tools: 6.12.1 -> 7.2.0
* pyproject_hooks: 1.0.0
* setuptools: 60.10.0 -> 68.0.0
* wheel: 0.38.4 -> 0.41.0
* zipp: 1.0.0 -> 3.16.2

Fixes #1351

---------

Co-authored-by: Greg <groodt@gmail.com>