internal(pystar): Make py_runtime_pair and autodetecting toolchain loadable.
They aren't quite usable yet. This mostly fixes references and adds
surrounding infrastructure to make this loadable and somewhat runnable.

Note that the "autodetecting" name is misleading: it doesn't really
autodetect anything. But it's the established name and part of the
public API.

The autodetecting toolchain is trimmed down from what Bazel does. The
Bazel version has a Python 2 variant and a "non strict" mode to support
that. With Python 2 no longer supported, that code is removed.

* Also alphabetically sorts the bzl_libraries in //python/private

Work towards #1069