tests(pystar): add analysis tests to cover basic windows building (#1452)
The CI currently only runs on Ubuntu, so it assumes that is the target
platform. This ends
up missing some notable Windows code paths, though.

Since its analysis-phase logic, we can force the platform to be Windows
for the analysis
tests, and then the rules follow the code paths that should be taken
under Windows.
This allows testing Windows logic under Ubuntu.