docs: move dependency management into respective bzl packages (#1459)
This moves the dependency management of what transitive files are needed
to generate docs for a .bzl file out of the docs directory and into the
respective bzl file's directory. This ensures that the bzl_library
targets we expose to users contain all the necessary dependencies.

Because there are some projects using the bzl_library targets in //docs,
some compatiblity aliases are added to make their migration path easier.
Those targets only public for historical reasons and shouldn't be used.

Work towards #1458