refactor(bzlmod): move bzlmod code to private/bzlmod (#1477)
This PR just moves all of the private `bzlmod` code to
`python/private/bzlmod`
and adds minimal `bzl_library` bindings to make the docs the same. Once
#1476
is merged, we can start exposing documentation for `module_extension`.

This includes extras in `pip_install/pip_repository.bzl` just to make it
possible to review and merge #1476 and this in parallel.