fix(repo setup): Skip aliases for unloaded toolchains (#1473)
Some platforms don't contain every version, e.g. s390x doesn't have 3.8,
which is
indicated by a missing sha256 value. When this happens, no repository
for the runtime
is created (`python_repository` rule).

Similar logic needs to be in the toolchains setup logic because
otherwise a reference
to an undefined repository exists in the select() expression of the
aliases. Because those
references are lazily evaluated, they don't always cause a problem, but
do mean that
query operations (e.g., `rdeps()`) can't work and the order of entries
is important
(which is surprising).

Closes #1472