test(bzlmod): Make some tests bzlmod compatible with Bazel@HEAD (#1482)
A few tests weren't compatible with bzlmod, so would fail when it was
enabled. The various causes and fixes are:

* Under bzlmod, `runfiles.CurrentRepository()` returns the empty string
for the main repository. To fix, an environment variable is used to tell
the test whether bzlmod is enabled or not.

* Accessing data files through `TEST_SRCDIR` directly is error-prone
under bzlmod because the directory name within runfiles changes from the
workspace name to `_main`. To fix, use the runfiles libraries, which
know how to map apparent repo names to the actual directory name. In the
integration tests, the runfiles library isn't available, so just check
for the `_main` directory instead.

Work towards #1469