refactor(visibility)!: limit visibility of an internal library (#1490)
This was arguably made too-visible previously and since all of the
consumable
symbols by the end users are re-exported via the //python:pip_bzl, we
can keep
everything that is in pip_install internal.

This could be a breaking change for people who are depending on internal
symbols or if they are using `whl_library` to create their own
implementation
of the `pip_repository` rule and are generating documentation
internally.
However, in that case they can apply a small patch to change the
visibility of
the `pip_repository_bzl` target.