feat(whlmaker): introduce an internal _WhlFile class and stop sorting RECORD (#1488)
This class is for being able to more easily recreate a wheel file after
extracting it. This is not intended for usage outside the rules_python
project. Also stop sorting the entries when writing a RECORD file making
the order of the RECORD file to be the same as the order the files to
the zip file are added.

Towards #1076