fix(gazelle): generate a single `py_test` for coarse_grained setups (#1538)
With a recent change to generate a single test target per python file
we introduced a regression for projects using
`# gazelle:python_generation_mode project` configuration directive
if there are multiple files with the same filenames but under different
directories within the source tree. This PR fixes the behaviour so that
we just generate a single target containing all test files as there is
no sure way to denormalize the paths so that they never clash.

Fixes #1442.