fix: always ignore `.pyc.NNNN` files from the hermetic runtime tree
Part of the pyc compilation process is to create a temporary file named
`<name>.pyc.NNNN`, where `NNNN` is a timestamp. Once the pyc is entirely
written, this file is renamed to the regular pyc file name. These files
only exist for brief periods of time, but its possible for different
threads/processes to see the temporary files when computing the glob()
values. Later, since the file is gone, an error is raised about the file
missing.

PR #1266 mostly fixed this issue, except that the glob exclude for an
interpreter runtime's files was behind the `ignore_root_user_error`
flag, which meant it wasn't always applied. This changes it to always be
applied, which should eliminate the failures.

Fixes #1261

Work towards #1520