test(pystar): run pystar under Windows and Mac
This is to have better test coverage. Only workspace for them is used
because of limited CI slots.

Also switches from last_green to last_rc, since a Bazel 7 RC is now
available. This will help catch any changes that need to be backported
from Bazel head to the Bazel 7 branch.

Work towards #1069