fix: py_proto_library: append to PYTHONPATH less (#1553)
Only append the `[proto_root]` to PYTHONPATH when it's actually
necessary, i.e. when the generated Python modules are under
_virtual_imports. Do no append it in the general case, when
`[proto_root]` is just the repository root.

This makes it much less likely that adding a transitive dep on a
`py_proto_library` will reorder the `PYTHONPATH`. Such reordering is
undesirable because it may lead to import errors.

Fixes #1551