feat(whl_library): set target_compatible_with based on whl name
This adds a pure starlark function to further parse the platform_tag
from the wheel and set the correct attributes for the
target generated by the whl_library rule. This should make errors
much more easy to understand when users try to build a docker image
using incompatible libraries.

This is currently hidden under an experimental flag described in the
changelog.