fix(bzlmod pip): ensure that sub-modules do not have invalid repos (#1549)
This fixes the cases where the 'default_version' is passed to the
'render_pkg_aliases' utility but the 'default_version' is not present
for the wheels. This usually happens when a sub-module is using the
'pip.parse' extension and the default_version can only be set by the
root module.

Previously, such a case would generate a `select()` expression that
mapped
the default condition to a non-existent target (because the sub-module
didn't
call `pip.parse()` with that version). This would either result in
errors
due the target not existing, or silently using a target intended for a
different Python version (which may work, but isn't correct to so).

Now, it results in a error via `select.no_match_error`.

Fixes #1548.