feat(gazelle): create py_binary targets for `if __name__ == "__main__"`
Previously a `py_binary` target was only created if a file named
`__main__.py` existed in the package. Now we look at each file that
would otherwise be added to a `py_library` target and see if it
contains a check for `if __name__ == "__main__"`. If such a check is
found in the file, then we create a `py_binary` target for it instead.

No existing tests fail with this change, though I think in real code
it's possible this could change target generation. New tests have been
added to check the new functionality.