feat: use rules_python implemented py_runtime, py_runtime_pair, PyRuntimeInfo (#1574)
This switches over to using the rules_python implementation of
`py_runtime`,
`py_runtime_pair`, and `PyRuntimeInfo` for Bazel 6 and higher. Bazel 5
lacks features
(specifically provider ctors) to allow enabling it on that version.

This is possible because the rules don't directly use the PyRuntimeInfo
provider
(mostly, see below), they only care about the structure of it as exposed
from the
ToolchainInfo provider.

Switching the toolchain providers and rules over early allows some
development of
the toolchain prior to Bazel 7 and the rest of the rules_python Starlark
implementation
being enabled.

The builtin PyRuntimeInfo is still returned and accepted for two
reasons:
* Better compatibility with the builtin rules to make transitioning
easier
* `py_binary` has an old, possibly defunct (not sure) code path that
will look up the
  the PyRuntimeInfo from a flag/implicit attribute.