fix(pip.parse): allow absolute path for python_interpreter; skip hermetic toolchain lookup when used (#1619)
The `python_interpreter` arg wasn't being properly handled in the bzlmod
code in two ways.

1. Lookup of a hermetic runtime wasn't being skipped when it was set.
The net effect was it
ignored the specified interpreter and would try to lookup a hermetic
interpreter using
the python version. To fix, add a check for python_interpreter to the
guard of the lookup.

2. Specifying an absolute path for the value wasn't being converted to a
`path` object,
which meant a plain string eventually made its way to some code
expecting a `path`
object. To fix, call `repository_ctx.path()` on the path to convert it
to a path object.

Fixes https://github.com/bazelbuild/rules_python/issues/1618