fix(whl_library): correctly handle osx universal2 wheel platform_tag parsing
It seems that the CI did not catch this and I am wondering if we run our
bzlmod example on a Mac. In the future I think that this code may need
to move to starlark, because the same parsing from wheel name to
platform identifiers is needed in #1685, but I'll do that in a separate
PR.

No changelog is needed because the feature this is fixing is unreleased
yet.