refactor: make the algorithm more conceptually correct
Use the fact that the selects are chosen based on the specificity
of the condition and also make the manipulation of the deps correct
through iteration of each Requires-Dist so that we don't need to
do any post-processing in the build() function. This should hopefully
make the extension to supporting python version, interpreter implementation
and others easier as we are generalizing what the Platform and Deps
data structures can do.

I have added a test to ensure coverage of the new code.