Allow using the current bazel version.
This should make it possible for the bazel-in-bazel tests to use
the same version of bazel as the outer invocation. This is important
so that Bazel CI's downstream testing pipeline is able to run tests
with unreleased versions of bazel automatically.

Along the way, fix an issue that blows up memory and cpu usage:
symlinks in the child workspace were being followed, which caused
the set of sources passed to the bazel-in-bazel test to become
30,000+ entries long. To avoid that, cap the size of the glob and
add those symlinks to the ignore files.

Also add a small optimization for how the workspace files are collected:
the bazel_integration_tests macro creates a separate filegroup target
for the list of files passed into it for each version of bazel it
tests, which duplicates that same list. To avoid that duplication,
manually create a filegroup once and pass that single list into
the bazel_integration_tests macro.