feat(py_runtime): Allow `py_runtime` to take an executable target as the interpreter (#1621)
This PR allows `py_runtime` to accept an executable (e.g. `sh_binary`).

This makes it easier to customize the interpreter binary used, as it
allows
intercepting invocation of the interpreter. For example, it can be used
to
change how the interpreter searches for dynamic libraries.

Related to https://github.com/bazelbuild/rules_python/issues/1612

---------

Co-authored-by: Richard Levasseur <rlevasseur@google.com>